+++ src/lw_common/ui/format/column_formatters/cell.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 
 namespace lw_common.ui.format.column_formatters {
     // formats the whole cell
-    class cell : column_formatter {
+    class cell : column_formatter_base {
         // extra formatting, if any
         private text_part formatting_ = null;