+++ src/lw_common/ui/format/column_formatters/format.cs
@@ -16,7 +16,7 @@ namespace lw_common.ui.format.column_formatters {
     - allow for date/time as well (color_date/color_time)
 
     */
-    class format : column_formatter {
+    class format : column_formatter_base {
         private multiline multi_ = new multiline();
         private cell color_ = new cell();
         private color_date date_ = new color_date();
@@ -25,7 +25,7 @@ namespace lw_common.ui.format.column_formatters {
         private format_number format_number_ = new format_number();
         private alternate_bg_color alternate_bg_ = new alternate_bg_color();
 
-        private List<column_formatter> sub_ = new List<column_formatter>(); 
+        private List<column_formatter_base> sub_ = new List<column_formatter_base>(); 
 
 
         public format() {
@@ -36,6 +36,16 @@ namespace lw_common.ui.format.column_formatters {
             sub_.Add(compare_number_);
             sub_.Add(format_number_);
             sub_.Add(alternate_bg_);
+        }
+
+        internal override void toggle_number_base() {
+            format_number_.toggle_number_base();
+        }
+
+        internal override void toggle_abbreviation() {
+            foreach ( var formatter in sub_)
+                if ( formatter is abbreviation)
+                    formatter.toggle_abbreviation();
         }
 
         internal override void load_syntax(settings_as_string sett, ref string error) {
@@ -58,9 +68,19 @@ namespace lw_common.ui.format.column_formatters {
                     sub_.Add(regex);
                 } else break;
             }
+
+            for (int idx = 0;; idx++) {
+                string prefix = "abb" + (idx > 0 ? "" + (idx + 1) : "");
+                string abbr_expr = sett.get(prefix + ".expr");
+                if (abbr_expr != "") {
+                    var abb = new abbreviation();
+                    load_sub_syntax(sett, abb, prefix, ref error);
+                    sub_.Add(abb);
+                } else break;
+            }
         }
 
-        private void load_sub_syntax(settings_as_string sett, column_formatter sub, string prefix, ref string error) {
+        private void load_sub_syntax(settings_as_string sett, column_formatter_base sub, string prefix, ref string error) {
             prefix += ".";
             settings_as_string sub_sett = new settings_as_string("");
             foreach ( var name in sett.names())