+++ src/lw_common/ui/format/column_formatters/helper/abbreviation.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace lw_common.ui.format.column_formatters.helper {
+    class abbreviation : regex_color {
+
+
+        internal override void toggle_abbreviation() {
+            
+        }
+    }
+}